refactor(Lotto Winning Number): 파일의 위치를 객체지향적으로 옮겨라
바꾼 이유:
- controller를 넣어서 해당 기능들을 모두 실행해야된다
- 차라리 그럴거면 lottoWinningNumber라는 패키지안에 넣자라는 생각
